[refactor] [move] FE events 구조화 변경 - 파일 구조에 맞게 위치 이동 REF : #11
