+++ web/src/app/page/home/section/HighlightSection.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "@/app/page/home/Home.module.scss";
 import { Link } from "react-router-dom";
-import EventSwiper from "@/domains/events/feature/components/EventSwiper/EventSwiper.jsx";
+import EventSwiper from "@/domains/events/feature/EventSwiper.jsx";
 import CouponSlideSection from "@/domains/coupons/feature/public/components/slide/CouponSlideSection.jsx";
 
 const HighlightSection = ({title, link, type}) => {
