+++ web/src/domains/events/feature/admin/components/table/EventTable.module.scss
