+++ web/src/domains/events/feature/admin/hooks/useEventEdit.js
@@ -1,6 +1,6 @@
 import { useDispatch } from "react-redux";
 import { requestUpdateExternalEvent } from "@/domains/events/store/index.js";
-import { useEventForm } from "./internal/useEventForm.js";
+import { useEventForm } from "./useEventForm.js";
 
 export const useEventEdit = ({ event, onSuccess }) => {
   const dispatch = useDispatch();
