+++ web/src/domains/events/feature/admin/pages/AdminEventPage.jsx
@@ -1,0 +1,7 @@
+import EventListAdmin from "@/domains/events/feature/admin/components/EventListAdmin.jsx";
+
+const AdminEventPage = () => {
+  return <EventListAdmin />
+};
+
+export default AdminEventPage;
