--- web/src/domains/events/feature/index.js
@@ -1,5 +1,0 @@
-// export let AdminEventList;
-
-// feature/event/index.js
-export { default as AdminEventList } from "./components/admin/list/EventListAdmin.jsx";
-export { useEventListAdmin } from "./hooks/admin/useEventListAdmin.js";
