+++ web/src/domains/events/feature/public/pages/UserEventPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import UserEventList from "@/domains/events/feature/components/user/lists/UserEventList.jsx";
+import UserEventList from "@/domains/events/feature/public/components/UserEventList.jsx";
 
 const UserEventPage = () => {
   return <UserEventList  />
