--- web/src/domains/events/page/admin/AdminEventPage.jsx
@@ -1,7 +1,0 @@
-import { AdminEventList } from "@/domains/events/feature/index.js";
-
-const AdminEventPage = () => {
-  return <AdminEventList />
-};
-
-export default AdminEventPage;
